Rename SUSHIControls -> FSHControls and RunSUSHI -> FSHHelpers
